ajout sous_total dynamique

+++ database/migrations/2026_04_10_130206_create_devis_materiel_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->foreignId('materiel_id');
             $table->integer('sous_devis')->default(1);
             $table->integer('quantite')->default(1);
-            $table->integer('prix')->nullable();
-            $table->integer('tva')->nullable();
+            $table->decimal('prix', 12, 2)->nullable();
+            $table->decimal('tva', 12, 2)->nullable();
 
             $table->timestamps();
             $table->foreign('devis_id')->references('id')->on('devis')->cascadeOnDelete();
